requirements.txt
streamlit 
yfinance 
pandas 
numpy 
plotly 
pytrends 
Pillow 
